fix nyiso
