+++ gridstatus/tests/base_test_iso.py
@@ -2,6 +2,7 @@ import pandas as pd
 import pytest
 from pandas.core.dtypes.common import is_numeric_dtype
 
+import gridstatus
 from gridstatus.base import GridStatus, _interconnection_columns
 
 
@@ -78,6 +79,31 @@ class BaseTestISO:
 
         # make sure right number of days are returned
         assert df["Time"].dt.day.nunique() == num_days
+
+    def test_range_two_days_with_day_start_endpoint(self):
+        yesterday = gridstatus.utils._handle_date(
+            "today",
+            self.iso.default_timezone,
+        ) - pd.Timedelta(days=1)
+        yesterday = yesterday.replace(hour=0, minute=5, second=0, microsecond=0)
+        start = yesterday - pd.Timedelta(minutes=10)
+        df = self.iso.get_fuel_mix(start=start, end=yesterday)
+
+        assert df["Time"].max() >= yesterday
+        assert df["Time"].min() <= start
+
+        self._check_fuel_mix(df)
+
+    def test_start_end_same_day(self):
+        yesterday = gridstatus.utils._handle_date(
+            "today",
+            self.iso.default_timezone,
+        ) - pd.Timedelta(days=1)
+        start = yesterday.replace(hour=0, minute=5, second=0, microsecond=0)
+        end = yesterday.replace(hour=6, minute=5, second=0, microsecond=0)
+        df = self.iso.get_fuel_mix(start=start, end=end)
+        assert df["Time"].dt.date.unique().tolist() == [yesterday.date()]
+        self._check_fuel_mix(df)
 
     def test_get_fuel_mix_latest(self):
         df = self.iso.get_fuel_mix("latest")
